divergence, still need pressure etc

+++ app/src/main/res/raw/divergence_frag.glsl
@@ -11,9 +11,9 @@ void main()
 
     // Find neighboring velocities:
     vec2 vN = texture2D(textureUnit0, T +vec2(0, 1)).xy;
-    vec2 vS = texture2D(textureUnit0, T +ivec2(0, -1)).xy;
-    vec2 vE = texture2D(textureUnit0, T +ivec2(1, 0)).xy;
-    vec2 vW = texture2D(textureUnit0, T + ivec2(-1, 0)).xy;
+    vec2 vS = texture2D(textureUnit0, T +vec2(0, -1)).xy;
+    vec2 vE = texture2D(textureUnit0, T +vec2(1, 0)).xy;
+    vec2 vW = texture2D(textureUnit0, T +vec2(-1, 0)).xy;
 
     // Find neighboring obstacles:
   //  vec3 oN = texelFetchOffset(textureUnit1, T, 0, ivec2(0, 1)).xyz;
